Refactor: Move Collector into http.rs
