+++ photon/src/template.rs
@@ -4,11 +4,11 @@ use std::{cell::RefCell, rc::Rc};
 use crate::{
     cache::{Cache, RegexCache},
     get_config,
-    http::{bake_ctx, get_bracket_pattern, HttpReq, HttpResponse},
+    http::{bake_ctx, get_bracket_pattern, Collector, CurlHandle, HttpReq, HttpResponse},
     template_executor::ExecutionOptions,
     PhotonContext,
 };
-use curl::easy::{Easy2, Handler, WriteError};
+use curl::easy::Easy2;
 use photon_dsl::{
     dsl::{CompiledExpression, Value, VariableContainer},
     parser::compile_expression_validated,
@@ -306,7 +306,7 @@ impl Matcher {
         context: &Context,
         photon_ctx: &PhotonContext,
     ) -> bool {
-        if let MatcherType::Status(_) = self.r#type {
+        if matches!(self.r#type, MatcherType::Status(_)) {
             return self.matches_status(data.status_code);
         }
 
@@ -583,7 +583,7 @@ impl HttpRequest {
         req: &HttpReq,
         base_url: &str,
         options: &ExecutionOptions,
-        curl: &mut Easy2<Collector>,
+        curl: &mut CurlHandle,
         ctx: &mut Context,
         photon_ctx: &PhotonContext,
         req_counter: &mut u32,
@@ -615,7 +615,7 @@ impl HttpRequest {
         &self,
         base_url: &str,
         options: &ExecutionOptions,
-        curl: &mut Easy2<Collector>,
+        curl: &mut CurlHandle,
         regex_cache: &RegexCache,
         parent_ctx: Rc<RefCell<Context>>,
         photon_ctx: &PhotonContext,
@@ -647,30 +647,6 @@ impl HttpRequest {
     }
 }
 
-pub struct Collector(pub Vec<u8>, pub Vec<u8>);
-
-impl Handler for Collector {
-    fn write(&mut self, data: &[u8]) -> Result<usize, WriteError> {
-        self.0.extend_from_slice(data);
-        Ok(data.len())
-    }
-
-    fn header(&mut self, data: &[u8]) -> bool {
-        // Make sure we're appending headers only, curl also gives us the HTTP response code header as well for some reason
-        if data.contains(&b':') {
-            self.1.extend_from_slice(data);
-        }
-        true
-    }
-}
-
-impl Collector {
-    pub fn reset(&mut self) {
-        self.0.clear();
-        self.1.clear();
-    }
-}
-
 impl Template {
     // TODO: Look into reducing the number of parameters
     // e.g. Refactor TemplateExecutor to forward some ExecutorContext with execution info like options, context, caches, request counter etc
@@ -679,7 +655,7 @@ impl Template {
         &self,
         base_url: &str,
         options: &ExecutionOptions,
-        curl: &mut Easy2<Collector>,
+        curl: &mut CurlHandle,
         parent_ctx: Rc<RefCell<Context>>,
         photon_ctx: &PhotonContext, // TODO: we can move parent_ctx into here, options as well
         req_counter: &mut u32,
