+++ photon/src/template_executor.rs
@@ -9,8 +9,10 @@ use url::Url;
 
 use crate::{
     cache::{Cache, RegexCache},
-    get_config, init_functions,
-    template::{Collector, Context, Template},
+    get_config,
+    http::Collector,
+    init_functions,
+    template::{Context, Template},
     template_loader::TemplateLoader,
     PhotonContext,
 };
